+++ imerald.cpp
@@ -1,6 +1,6 @@
 #include "imerald.h"
 
-Imerald::Imerald(qreal pos_x, qreal pos_y) : GraphicPixmapObject(pos_x, pos_y, "Imerald.png")
+Imerald::Imerald(int pos_x, int pos_y) : GraphicPixmapObject(pos_x, pos_y, "Imerald.png")
 {
     sizeOfPictureX = 8;
     sizeOfPictureY = 7;
